+++ lib/src/flutter_facebook_connect_base.dart
@@ -9,24 +9,44 @@ import 'package:http/http.dart' as http;
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+const String _kFacebookConnectTokenKey = "FACEBOOK_CONNECT_TOKEN_KEY";
+
 class FacebookConnect {
-  final String _facebookConnectTokenKey = "FACEBOOK_CONNECT_TOKEN_KEY";
-  final FacebookOptions options;
-  final String responseContent;
-  final bool fullscreen;
-  final bool storeToken;
+  final String appId;
+  final String clientSecret;
+  final String version;
   final flutterWebviewPlugin = new FlutterWebviewPlugin();
 
-  FacebookConnect(this.options,
-      {this.responseContent = "", this.fullscreen = false, this.storeToken = true});
+  static FacebookConnect _instance;
 
-  Future<FacebookOAuthToken> login({bool force = false}) async {
-    if (force == false && storeToken) {
-      _token = await _getStoredToken();
-    }
+  FacebookOAuthToken get token => _token;
 
+  FacebookConnect._({this.appId, this.clientSecret, this.version});
+
+  factory FacebookConnect(
+      {@required String appId,
+      @required String clientSecret,
+      String version = 'v2.10'}) {
+    return _instance ??= new FacebookConnect._(
+        appId: appId, clientSecret: clientSecret, version: version);
+  }
+
+  /// Log user to Facebook using a webview
+  ///
+  /// [scope] see available scope in [FacebookAuthScope]
+  ///   - more info [here](https://developers.facebook.com/docs/facebook-login/permissions/)
+  /// [fullscreen] launch the webview in fullscreen or not on iOS
+  /// [storeToken] store [FacebookOAuthToken] or not inside [SharedPreferences]
+  Future<FacebookOAuthToken> login({
+    bool force = false,
+    List<String> scope,
+    bool cookie = true,
+    bool fullscreen = false,
+    bool storeToken = true,
+  }) async {
+    _token = await getStoredToken();
     if (_shouldRequestCode(force: force)) {
-      revoke();
+      logout();
       // close any open browser (happen on hot reload)
       await flutterWebviewPlugin.close();
       _isOpen = true;
@@ -44,36 +64,44 @@ class FacebookConnect {
         _close();
       });
 
-      String url = "https://www.facebook.com/dialog/oauth?client_id=${options
-          .appId}&redirect_uri=http://localhost:8080/";
-      if (options.scope != null) {
-        url += "&scope=${options
-            .scope}";
+      String url =
+          "https://www.facebook.com/dialog/oauth?client_id=${appId}&redirect_uri=http://localhost:8080/";
+      if (scope?.isNotEmpty == true) {
+        url += "&scope=${scope.join(",")}";
       }
 
-
       // launch url inside webview
-      flutterWebviewPlugin.launch(
-          url,
-          clearCookies: !options.cookie,
-          fullScreen: fullscreen);
+      flutterWebviewPlugin.launch(url,
+          clearCookies: !cookie, fullScreen: fullscreen);
 
       _code = await _onCode.first;
-      _close();
-      _token = await _getToken();
+      _token = await _requestToken();
       if (storeToken) {
         _storeToken();
       }
+      _close();
     }
     return _token;
   }
 
-  void revoke() {
+  void logout() {
     _token = null;
-    if (storeToken) {
-      _storeToken();
-    }
+    _storeToken();
     _code = null;
+  }
+
+  /// Get the [FacebookOAuthToken] stored inside [SharedPreferences]
+  static Future<FacebookOAuthToken> getStoredToken() async {
+    try {
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      String token = prefs.getString(_kFacebookConnectTokenKey);
+      if (token != null && token.isNotEmpty) {
+        return new FacebookOAuthToken.fromMap(JSON.decode(token));
+      }
+    } catch (e) {
+      print(e);
+    }
+    return null;
   }
 
   ///////////
@@ -87,35 +115,9 @@ class FacebookConnect {
   HttpServer _server;
   Stream<String> _onCodeStream;
 
-  Future<FacebookOAuthToken> _getStoredToken() async {
-    try {
-      SharedPreferences prefs = await SharedPreferences.getInstance();
-      String token = prefs.getString(_facebookConnectTokenKey);
-      if (token != null && token.isNotEmpty) {
-        return new FacebookOAuthToken.fromMap(JSON.decode(token));
-      }
-    } catch (e) {
-      print(e);
-    }
-    return null;
-  }
-
-  Future _storeToken() async {
-    try {
-      SharedPreferences prefs = await SharedPreferences.getInstance();
-      prefs.setString(_facebookConnectTokenKey, _token != null ? JSON.encode(_token.toMap()) : null);
-      await prefs.commit();
-    } catch (e) {
-      print(e);
-    }
-  }
-
-  Future<FacebookOAuthToken> _getToken() async {
+  Future<FacebookOAuthToken> _requestToken() async {
     final http.Response response = await http.get(
-        "https://graph.facebook.com/${options
-            .version}/oauth/access_token?client_id=${options
-            .appId}&redirect_uri=http://localhost:8080/&client_secret=${options
-            .clientSecret}&code=$_code");
+        "https://graph.facebook.com/${version}/oauth/access_token?client_id=${appId}&redirect_uri=http://localhost:8080/&client_secret=${clientSecret}&code=$_code");
 
     if (response.statusCode >= 200 && response.statusCode < 300) {
       return new FacebookOAuthToken.fromMap(JSON.decode(response.body));
@@ -124,7 +126,19 @@ class FacebookConnect {
     return null;
   }
 
-  bool _shouldRequestCode({bool force = false}) => force || _token == null;
+  Future _storeToken() async {
+    try {
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      prefs.setString(_kFacebookConnectTokenKey,
+          _token != null ? JSON.encode(_token.toMap()) : null);
+      await prefs.commit();
+    } catch (e) {
+      print(e);
+    }
+  }
+
+  bool _shouldRequestCode({bool force = false}) =>
+      _token == null || force == true;
 
   Stream<String> get _onCode =>
       _onCodeStream ??= _onCodeCtrl.stream.asBroadcastStream();
@@ -153,7 +167,7 @@ class FacebookConnect {
       request.response
         ..statusCode = 200
         ..headers.set("Content-Type", ContentType.HTML.mimeType)
-        ..write(responseContent);
+        ..write("");
 
       final String code = request.uri.queryParameters["code"];
       final error = uri.queryParameters["error"];
@@ -166,19 +180,6 @@ class FacebookConnect {
       }
     });
   }
-}
-
-class FacebookOptions {
-  final String appId;
-  final String scope;
-  final bool cookie;
-  final String clientSecret;
-
-  //final bool xfbml;
-  final String version;
-
-  const FacebookOptions(
-      {@required this.appId, @required this.clientSecret, this.version = 'v2.10', this.scope, this.cookie = false});
 }
 
 class FacebookOAuthToken {
@@ -195,4 +196,58 @@ class FacebookOAuthToken {
 
   Map toMap() =>
       {'access_token': access, 'token_type': type, 'expires_in': expiresIn};
+}
+
+class FacebookAuthScope {
+  static String get publicProfile => "public_profile";
+
+  static String get userFriends => "user_friends";
+  static String get email => "email";
+  static String get userAboutMe => "user_about_me";
+  static String get userActionsBooks => "user_actions.books";
+  static String get userActionsFitness => "user_actions.fitness";
+  static String get userActionsMusic => "user_actions.music ";
+  static String get userActionsNews => "user_actions.news";
+  static String get userActionsVideo => "user_actions.video";
+  static String get userBirthday => "user_birthday";
+  static String get userEducationHistory => "user_education_history";
+  static String get userEvents => "user_events";
+  static String get userGamesActivity => "user_games_activity";
+  static String get userHometown => "user_hometown";
+  static String get userLikes => "user_likes";
+  static String get userLocation => "user_location";
+  static String get userManagedGroups => "user_managed_groups";
+  static String get userPhotos => "user_photos";
+  static String get userPosts => "user_posts";
+  static String get userRelationships => "user_relationships";
+  static String get userRelationshipDetails => "user_relationship_details";
+  static String get userReligion_politics => "user_religion_politics";
+  static String get userTaggedPlaces => "user_tagged_places";
+  static String get userVideos => "user_videos";
+  static String get userWebsite => "user_website";
+  static String get userWork_history => "user_work_history";
+  static String get readCustomFriendlists => "read_custom_friendlists";
+  static String get readInsights => "read_insights";
+  static String get readAudienceNetworkInsights =>
+      "read_audience_network_insights";
+  static String get readPageMailboxes => "read_page_mailboxes";
+  static String get managePages => "manage_pages";
+  static String get publishPages => "publish_pages";
+  static String get publishActions => "publish_actions";
+  static String get rsvpEvent => "rsvp_event";
+  static String get pagesShowList => "pages_show_list";
+  static String get pagesManageCta => "pages_manage_cta";
+  static String get pagesManageInstantArticles =>
+      "pages_manage_instant_articles";
+  static String get adsRead => "ads_read";
+  static String get adsManagement => "ads_management";
+  static String get businessManagement => "business_management";
+  static String get pagesMessaging => "pages_messaging";
+  static String get pagesMessagingSubscriptions =>
+      "pages_messaging_subscriptions";
+  static String get pagesMessagingPayments => "pages_messaging_payments";
+  static String get pagesMessagingPhoneNumber => "pages_messaging_phone_number";
+
+  static String userActions(String appNamespace) =>
+      "user_actions:$appNamespace";
 }